Veloviewer Map Flipper second try
I tried a bit different aproach here and hope it works this way. Note that in order for MapFlipper to work, you have to select "Satellite Map" in the map itself.
There might be some minor mistakes in my code, because I can't test it, but I'm confident that this concept might work.