Avoid clipping of goal marker text
Previously, when the marker gets near the end of the progress bar the
marker text would be clipped. Now we switch the `text-anchor` to avoid
this. Note that this introduces the awkwrad `GOAL_MARKER_TEXT_WIDTH`
constant which is the approximate width of the text marker. This is
necessary because we can't determine the 'real' width of the marker text
without allowing it to be rendered first.