Incorporate ramp rate calculations into Elevate
To complete this feature, the dayFitnessTrend model and fitness service were changed to calculate fitness ramp rates for 7 days, 28 days, 90 days, and 365 days in the past. The fitness-trend-legend component was modified to display the ramp rates below the fitness, fatigue, and form values for the displayed day. The graph component and viewableFitnessData models were modified to include the plot for the weekly ramp rate (similar to how Golden Cheetah displays this plot). The Elevate themes were modified to incorporate a third color (the primary and secondary palletes did not give much room for a different enough color) as the warn color (I saw some references to using the angular warn color, but one was never set, so I don't think this will affect anything). Finally, a description of how to interpret ramp rate and links to resources was added to the fitness-info-dialog component.