Fix stress score discrepancy
The discrepancy in stress score values between the Fitness Trend and Activities pages is due to the use of floor() in Fitness Trend and round() in Activities. I'm sure when first creating the program there was some reasoning behind using floor() on the Fitness Trend page, but changing to round() was the easiest solution of the three:
1) Modify the stress calculation at the source (activity-computer) to output a number with no decimals (rounded, or floored) - pretty simple but I don't think the after effects (losing precision for future calculations) would be worth it.
2) Modify the Activities table namespace to write a custom detection for stress scores to floor() the values instead of round() - pretty roundabout and involved due to how the namespace and table is programmed already
3) Modify the Fitness Trend and day stress model to round() instead of floor() - easiest and most practical: it passes all of the unit tests without modifying the tests

Ultimately, one has to weigh the decision between floor() and round() for these values, and I decided that round() would produce a more appropriate value. If people are looking at the stress scores to calculate weekly TSS to plan training around, 1 stress point here and there is still just a rounding issue (pun intended), but I feel as though _overestimating_ stress would result in less potential for overload than underestimating stress.